feat: added price fetching for Fipe Form
Body:
This commit primarily achieves two updates:

Firstly, the changes adjust the revalidation timing for API requests in the `brands`, `models`, and `years` routes. There was an initial misinterpretation of the units for setting revalidation time, where milliseconds were provided instead of seconds. This adjustment allows the cache revalidation to accurately occur every 1 hour as originally intended - enhancing efficiency and preventing unnecessary API calls.

Secondly, this commit introduces core functionality to the Fipe Form that allows for vehicle price fetching. Users are now able to receive real-time price feedback for a selected vehicle model. The price fetched is displayed in a new, styled component that enhances user experience and usability of the app.

Together, these changes boost application performance and functionality, contributing to a more engaging and informative user experience.