feat: Integrate state management in FipeForm for Price Display
Body:
Incorporated useState to improve the behavior and control of when price is shown in the Fipe form. The Price information display is now tied to a local 'showPrice' state, which is updated whenever brand or model changes and after the price fetch on form submission. This ensures the price is concealed during form adjustments and revealed accurately only once the user finishes their input and asks for the pricing information. This approach streamlines the user experience by preventing premature or confusing price display.