styles: Enhanced UX and Translated App to Portuguese
Body:
The underlying purpose of these changes is two-fold: improvise the user experience and localize the application to Portuguese. The revisions comprise alterations to the error messages, metadata, and language, thereby ensuring improved user interactions for Portuguese-speaking users.

Significant modifications involve the incorporation of a suspense fallback with a `<CircularProgress />` component to deliver progressive feedback in form submission, which can be highly beneficial in instances of slow network connections.

Lastly, the changes include amendments for visual aesthetics, such as the repositioning of the submit button inside the form and subtle margin adjustments, to further refine the user interface and foster a seamless user experience.